+++ snakemake/modules.py
@@ -3,6 +3,7 @@ __copyright__ = "Copyright 2021, Johannes Köster"
 __email__ = "johannes.koester@uni-due.de"
 __license__ = "MIT"
 
+from pathlib import Path
 import types
 import re
 
@@ -43,6 +44,7 @@ class ModuleInfo:
         config=None,
         skip_validation=False,
         replace_prefix=None,
+        prefix=None,
     ):
         self.workflow = workflow
         self.name = name
@@ -50,9 +52,30 @@ class ModuleInfo:
         self.meta_wrapper = meta_wrapper
         self.config = config
         self.skip_validation = skip_validation
-        self.replace_prefix = replace_prefix
 
-    def use_rules(self, rules=None, name_modifier=None, ruleinfo=None):
+        if prefix is not None:
+            if isinstance(prefix, Path):
+                prefix = str(prefix)
+            if not isinstance(prefix, str):
+                raise WorkflowError(
+                    "Prefix definition in module statement must be string or Path."
+                )
+            if replace_prefix is not None:
+                raise WorkflowError(
+                    "Module definition contains both prefix and replace_prefix. "
+                    "Only one at a time is allowed."
+                )
+
+        self.replace_prefix = replace_prefix
+        self.prefix = prefix
+
+    def use_rules(
+        self,
+        rules=None,
+        name_modifier=None,
+        ruleinfo=None,
+        skip_global_report_caption=False,
+    ):
         snakefile = self.get_snakefile()
         with WorkflowModifier(
             self.workflow,
@@ -60,15 +83,17 @@ class ModuleInfo:
             base_snakefile=snakefile,
             skip_configfile=self.config is not None,
             skip_validation=self.skip_validation,
+            skip_global_report_caption=skip_global_report_caption,
             rule_whitelist=self.get_rule_whitelist(rules),
             rulename_modifier=get_name_modifier_func(rules, name_modifier),
             ruleinfo_overwrite=ruleinfo,
             allow_rule_overwrite=True,
             namespace=self.name,
             replace_prefix=self.replace_prefix,
+            prefix=self.prefix,
             replace_wrapper_tag=self.get_wrapper_tag(),
         ):
-            self.workflow.include(snakefile, overwrite_first_rule=True)
+            self.workflow.include(snakefile, overwrite_default_target=True)
 
     def get_snakefile(self):
         if self.meta_wrapper:
@@ -111,11 +136,13 @@ class WorkflowModifier:
         base_snakefile=None,
         skip_configfile=False,
         skip_validation=False,
+        skip_global_report_caption=False,
         rulename_modifier=None,
         rule_whitelist=None,
         ruleinfo_overwrite=None,
         allow_rule_overwrite=False,
         replace_prefix=None,
+        prefix=None,
         replace_wrapper_tag=None,
         namespace=None,
     ):
@@ -131,10 +158,11 @@ class WorkflowModifier:
         self.skip_configfile = skip_configfile
         self.rulename_modifier = rulename_modifier
         self.skip_validation = skip_validation
+        self.skip_global_report_caption = skip_global_report_caption
         self.rule_whitelist = rule_whitelist
         self.ruleinfo_overwrite = ruleinfo_overwrite
         self.allow_rule_overwrite = allow_rule_overwrite
-        self.path_modifier = PathModifier(replace_prefix, workflow)
+        self.path_modifier = PathModifier(replace_prefix, prefix, workflow)
         self.replace_wrapper_tag = replace_wrapper_tag
         self.namespace = namespace