+++ snakemake/remote/gfal.py
@@ -151,6 +151,9 @@ class RemoteObject(AbstractRemoteObject):
         # disable all timeouts (file transfers can take a long time)
         self._gfal("copy", "-p", "-f", "-n", "4", "-t", "0", "-T", "0", source, target)
 
+    def remove(self):
+        self._gfal("rm", "-r", self.remote_file())
+
     @property
     def list(self):
         # TODO implement listing of remote files with patterns