+++ tests/tests.py
@@ -429,6 +429,17 @@ def test_upstream_conda():
     run(dpath("test_conda"), use_conda=True, conda_frontend="conda")
 
 
+@skip_on_windows
+def test_deploy_script():
+    run(dpath("test_deploy_script"), use_conda=True)
+
+
+@skip_on_windows
+def test_deploy_hashing():
+    tmpdir = run(dpath("test_deploy_hashing"), use_conda=True, cleanup=False)
+    assert len(next(os.walk(os.path.join(tmpdir, ".snakemake/conda")))[1]) == 2
+
+
 def test_conda_custom_prefix():
     run(
         dpath("test_conda_custom_prefix"),
@@ -585,7 +596,7 @@ def test_restartable_job_cmd_exit_1_one_restart():
 def test_restartable_job_qsub_exit_1():
     """Test the restartable job feature when qsub fails
 
-    The qsub in the sub directory will fail the first time and succeed the
+    The qsub in the subdirectory will fail the first time and succeed the
     second time.
     """
     # Even two consecutive times should fail as files are cleared
@@ -674,10 +685,10 @@ def test_gs_requester_pays(
     Parameters
     ----------
     requesting_project: Optional[str]
-        User project to bill for download. None will not provide project for
+        User project to bill for download. None will not provide the project for
         requester-pays as is the usual default
     requesting_url: str
-        URL of bucket to download. Default will match expected output, but is a
+        URL of the bucket to download. The default will match the expected output but is a
         bucket that doesn't require requester pays.
     """
     # create temporary config file
@@ -1273,6 +1284,16 @@ def test_modules_all():
     run(dpath("test_modules_all"), targets=["a"])
 
 
+@skip_on_windows
+def test_modules_prefix():
+    run(dpath("test_modules_prefix"), targets=["a"])
+
+
+@skip_on_windows
+def test_modules_peppy():
+    run(dpath("test_modules_peppy"), targets=["a"])
+
+
 def test_modules_specific():
     run(dpath("test_modules_specific"), targets=["test_a"])
 
@@ -1382,3 +1403,37 @@ def test_checkpoint_allowed_rules():
 @skip_on_windows
 def test_modules_ruledeps_inheritance():
     run(dpath("test_modules_ruledeps_inheritance"))
+
+
+@skip_on_windows
+def test_conda_named():
+    run(dpath("test_conda_named"), use_conda=True)
+
+
+@skip_on_windows
+def test_default_target():
+    run(dpath("test_default_target"))
+
+
+def test_cache_multioutput():
+    run(dpath("test_cache_multioutput"), shouldfail=True)
+
+
+@skip_on_windows
+def test_github_issue1384():
+    try:
+        tmpdir = run(dpath("test_github_issue1384"), cleanup=False)
+        shell(
+            """
+            cd {tmpdir}
+            python -m snakemake --generate-unit-tests
+            pytest -v .tests/unit
+            """
+        )
+    finally:
+        shutil.rmtree(tmpdir)
+
+
+@skip_on_windows
+def test_peppy():
+    run(dpath("test_peppy"))